==> study5/source/04_bonus-pay.R <==
# load packages -----------------------------------------------------------

if(!require('pacman')) {
  install.packages('pacman')
}
pacman::p_load(tidyverse, here, hablar, psych, qualtRics, varhandle, magrittr)

# load data ---------------------------------------------------------------

clean <- read_csv(here::here("study5", "data", "clean.csv"))

# additional questions  ---------------------------------------------------


## main qs to be included & their answers: 
# Do you think men or women in this study correctly solved more multiplication problems on average?

## men tend to have higher scores on avg

## for questions about perc_gender_comp_F & perc_gender_comp_M, given study design, will just draw from prev lit
## so correct answer will be "Men would choose tournament more often than piece rate" (perc_gender_comp_M) & "Women would choose piece rate more often than tournament" (perc_gender_comp_F)

# Do you think men or women spent more time practicing/studying before multiplication task? 

## women practice the task more 

# Your task score was ${e://Field/task_score}. If my performance is compared to that of all participants that completed the task, I think my score will be...

clean <- clean %>% 
  mutate(percent_rank = percent_rank(task_score), conf_rank_decimal = clean$conf_rank/100) %>% drop_na(percent_rank)

clean$conf_accuracy <- 0


for (i in 1:nrow(clean)) {
  if (clean$percent_rank[i] >= clean$conf_rank_decimal[i]) 
  {clean$conf_accuracy[i] <- "correct"
  
  } else if (clean$percent_rank [i] < clean$conf_rank_decimal[i]) 
  {clean$conf_accuracy[i] <- "incorrect"
  }
}

# randomly selecting bonus question ---------------------------------------

## making sure to separate random selection based on condition bc ppts in practice condition has extra q

clean <- clean[!duplicated(clean$workerId),]
## shuffling order to make pairing up truly random 
clean <-clean[sample(1:nrow(clean), replace = F), ]
clean$bonus_assn <- 0

clean$bonus_assn <- rep(1:5, length.out = nrow(clean))

total_payment <- clean 

## conf_accuracy q == 1; perc_task_gender_pract == 2; better_gender_guess == 3; perc_gender_comp_M == 4; perc_gender_comp_F == 5
## "correct" ==1, "Women were more likely to practice/study than men" == 2; "Men" == 3; 
## "Men would choose tournament more often than piece rate" == 4; "Women would choose piece rate more often than tournament" == 5

total_payment$bonus_qs <- ifelse(total_payment$conf_accuracy == "correct" & total_payment$bonus_assn == 1, .1, 0)
total_payment$bonus_qs <- ifelse(total_payment$perc_task_gender_pract == "Women were more likely to practice/study than men" & total_payment$bonus_assn == 2, .1, total_payment$bonus_qs)
total_payment$bonus_qs <- ifelse(total_payment$better_gender_guess == "Men" & total_payment$bonus_assn == 3, .1, total_payment$bonus_qs)
total_payment$bonus_qs <- ifelse(total_payment$perc_gender_comp_M == "Men would choose tournament more often than piece rate" & total_payment$bonus_assn == 4, .1, total_payment$bonus_qs)
total_payment$bonus_qs <- ifelse(total_payment$perc_gender_comp_F == "Women would choose piece rate more often than tournament" & total_payment$bonus_assn == 5, .1, total_payment$bonus_qs)



# main task payoffs -------------------------------------------------------

clean %<>% drop_na(task_score)

### filtered based on piece rate assn
PR_choice <- subset(clean, condition == "piecerate", select = c(workerId, task_score))

## multiply by .10

PR_choice$PR_pay <- PR_choice$task_score*.10


### filtered based on choice for tournament
tourn_choice <- subset(clean, condition == "tournament", select = c(workerId, task_score))

#### randomly sort rows of DF

set.seed(100)
tourn_choice_even <- tourn_choice %>% slice_sample(n = nrow(tourn_choice)-1)


set.seed(100)
tourn_choice_even <-tourn_choice_even[sample(1:nrow(tourn_choice_even), replace = F), ]

## making dataset even by randomly selecting ppt to be removed (to be inserted & randomly matched with partner later)


first_half <- nrow(tourn_choice_even)/2
second_half_start <- nrow(tourn_choice_even)/2 + 1
second_half_end <- nrow(tourn_choice_even)



### split DF in half 
tourn_choice_split1 <- tourn_choice_even[1:first_half, c("workerId", "task_score")]
tourn_choice_split2 <- tourn_choice_even[second_half_start:second_half_end, c("workerId", "task_score")]

tourn_choice_split1$ID <- c(1:first_half)
tourn_choice_split2$ID <- c(1:first_half)

combined2 <- merge (tourn_choice_split1, tourn_choice_split2, by = "ID")

### subtract round scores to determine winner 

combined2$difference_score <- combined2$task_score.x - combined2$task_score.y

###ID winners and put back in main dataset

difference_score <- combined2$difference_score
difference_score2 <- rep (difference_score)


difference_score <- append(difference_score, difference_score2)
tourn_choice_even$difference_score <- difference_score

tourn_choice_even$W_L <- "NA"


for (i in 1:first_half) {
  if (tourn_choice_even$difference_score [i]> 0) {tourn_choice_even$W_L[i] <- "win"
  
  } else if (tourn_choice_even$difference_score [i]< 0)
  {tourn_choice_even$W_L[i] <- "lose"
  
  } else if (tourn_choice_even$difference_score [i] == 0) 
  {tourn_choice_even$W_L[i] <- "tie"}
  
}


for (i in second_half_start:second_half_end) {
  if (tourn_choice_even$difference_score [i]> 0) {tourn_choice_even$W_L[i] <- "lose"
  
  } else if (tourn_choice_even$difference_score [i]< 0)
  {tourn_choice_even$W_L[i] <- "win"
  
  } else if (tourn_choice_even$difference_score [i] == 0) 
  {tourn_choice_even$W_L[i] <- "tie"}
  
}


## not needed because there aren't any ties  
# tourn_choice_even <- tourn_choice_even[-sample(1:sum(tourn_choice_even$W_L == "tie"), size = sum(tourn_choice_even$W_L == "tie")/2, replace = F),]


## randomly select partner for odd one out 
OOO <- anti_join(tourn_choice, tourn_choice_even)

## randomly selecting partner for them
set.seed(100)
tourn_choice_even %>% slice_sample(n =1)
## they lose tournament: opponent ID = A19CA1G04F23NN; score = 13, do not need to add as row to tourn_winners dataset

tourn_winners <- tourn_choice_even %>% filter(W_L == "win"| W_L == "tie") 
tourn_winners$tourn_payment <- 0

tourn_winners$tourn_payment <- tourn_winners$task_score*.20

##  Merging based on worker ID:

total_payment <- left_join(total_payment, PR_choice,by = "workerId")
total_payment <-left_join(total_payment, tourn_winners, by = "workerId")

## merging into one column
total_payment$combined_pay <- coalesce(total_payment$tourn_payment, total_payment$PR_pay)
## inserting $0 payment for people who lose tournaments

total_payment$combined_pay <- ifelse(is.na(total_payment$combined_pay), 0, total_payment$combined_pay)

tourn_selected <- total_payment %>% filter(condition == "tournament" & combined_pay > 0) %>% sample_n(250) 
PR <- total_payment %>% filter(condition == "tournament")

total_payment_final <- rbind(tourn_selected, PR) %>% filter(combined_pay > 0) %>% select(workerId, combined_pay)

write_csv(total_payment_final,here("study5", "data", 'total_payment_output.csv'), col_names = F)


## estimating total bonus costs including MTurk fees: 

sum(total_payment_final$combined_pay)*.20 + sum(total_payment_final$combined_pay)


# splitting workers into groups for cloudresearch  ------------------------

pay <- read_csv(here::here("study5", "data", 'total_payment_output.csv'), col_names = F)


names(pay) <- c("AmazonIdentifier", "payment")

## have to create separate datasets for each round of running the study 

studydata_full1 <- read_csv(here::here("study5", "data", 'studydata_full1.csv'))
studydata_full3 <- read_csv(here::here("study5", "data", 'studydata_full3.csv'))
studydata_pilot <- read_csv(here::here("study5", "data", 'studydata_pilot.csv'))


## need to match on the MTurk ID 

studydata_pilot_pay <- match_df(pay, studydata_pilot, on="AmazonIdentifier")

studydata_full1_pay <- match_df(pay, studydata_full1, on="AmazonIdentifier")
  
studydata_full3_pay <- match_df(pay, studydata_full3, on="AmazonIdentifier")

## then create separate datasets for each specific round 

write_csv(studydata_pilot_pay %>% unique(.),here::here("study5", "data", 'studydata_pilot_pay.csv'), col_names = F)

write_csv(studydata_full1_pay%>% unique(.),here::here("study5", "data", 'studydata_full1_pay.csv'), col_names = F)

write_csv(studydata_full3_pay %>% unique(.),here::here("study5", "data", 'studydata_full3_pay.csv'), col_names = F)



combined_subset <- rbind(studydata_pilot_pay, studydata_full1_pay, studydata_full3_pay)








pay %>% slice(1:100) %>% write_csv(.,here("study4", "data", 'first.csv'), col_names = F)
pay %>% slice(101:200) %>% write_csv(.,here("study4", "data", 'second.csv'), col_names = F)
pay %>% slice(201:300) %>% write_csv(.,here("study4", "data", 'third.csv'), col_names = F)
pay %>% slice(301:400) %>% write_csv(.,here("study4", "data", 'fourth.csv'), col_names = F)
pay %>% slice(401:500) %>% write_csv(.,here("study4", "data", 'fifth.csv'), col_names = F)
pay %>% slice(501:600) %>% write_csv(.,here("study4", "data", 'sixth.csv'), col_names = F)
pay %>% slice(601:700) %>% write_csv(.,here("study4", "data", 'seventh.csv'), col_names = F)
pay %>% slice(701:800) %>% write_csv(.,here("study4", "data", 'eighth.csv'), col_names = F)
pay %>% slice(801:900) %>% write_csv(.,here("study4", "data", 'ninth.csv'), col_names = F)
pay %>% slice(901:1000) %>% write_csv(.,here("study4", "data", 'tenth.csv'), col_names = F)
pay %>% slice(1001:1027) %>% write_csv(.,here("study4", "data", 'eleventh.csv'), col_names = F)
