==> study5/source/05_dataset-summary.Rmd <==
---
title: "prelim_results"
author: "Keana Richards"
output:
  html_document:
    chunk_output_type: console
    number_sections: yes
    theme: cerulean
    toc: yes
    toc_depth: 4
    toc_float: yes
  pdf_document:
    toc: yes
    toc_depth: '4'
editor_options: 
  chunk_output_type: console
---


```{r, include = F}
knitr::opts_chunk$set(echo = FALSE, message=FALSE, warning = F)


## limiting number of digits to 2 decimal places
options(digits = 2)
options(repos="https://cran.rstudio.com" )

# load packages -----------------------------------------------------------

if(!require('pacman')) {
  install.packages('pacman')
  }
pacman::p_load(tidyverse, here, hablar, psych, qualtRics, summarytools, varhandle, sjPlot, papaja)


## Install the stable development versions from GitHub
# if(!"papaja" %in% rownames(installed.packages())) devtools::install_github("crsh/papaja")
# if(!"emo" %in% rownames(installed.packages())) devtools::install_github("hadley/emo")
# library(emo)
# library(papaja)

clean <- read_csv(here("study5", "data", "clean.csv"))


# na_drop <- clean %>% drop_na(condition) 
extrafont::loadfonts(device="win")


options(scipen=999)
par(mar = c(2,2,2,2))


```


# Summary of dataset 


```{r, results = "asis"}
dfSummary(clean,  plain.ascii = FALSE, style = 'grid', graph.magnif = 0.75, 
          valid.col = FALSE, tmp.img.dir = "/tmp")
```


```{r}
inspect.na(clean)

```
